+++ steam_web/Script/DTO/ItemRender/ItemRenderListingAsset.cs
@@ -7,7 +7,7 @@ public class ItemRenderListingAsset
 {
     [JsonPropertyName("currency")] public byte Currency { get; init; }
     [JsonPropertyName("appid")] public uint AppId { get; init; }
-    [JsonPropertyName("contextid")] public uint ContextId { get; init; }
+    [JsonPropertyName("contextid")] public ulong ContextId { get; init; }
     [JsonPropertyName("id")] public ulong Id { get; init; }
     [JsonPropertyName("amount")] public ushort Amount { get; init; }
     [JsonPropertyName("market_actions")] public ItemAction[] MarketActions { get; init; } = Array.Empty<ItemAction>();
